+++ components/streams/src/stream.cpp
@@ -7,27 +7,27 @@ using ftl::stream::Stream;
 
 const ftl::codecs::Channels<0> &Stream::available(int fs) const {
 	SHARED_LOCK(mtx_, lk);
-	if (fs < 0 || fs >= state_.size()) throw ftl::exception("Frameset index out-of-bounds");
+	if (fs < 0 || static_cast<uint32_t>(fs) >= state_.size()) throw ftl::exception("Frameset index out-of-bounds");
 	return state_[fs].available;
 }
 
 const ftl::codecs::Channels<0> &Stream::selected(int fs) const {
 	SHARED_LOCK(mtx_, lk);
-	if (fs < 0 || fs >= state_.size()) throw ftl::exception("Frameset index out-of-bounds");
+	if (fs < 0 || static_cast<uint32_t>(fs) >= state_.size()) throw ftl::exception("Frameset index out-of-bounds");
 	return state_[fs].selected;
 }
 
 void Stream::select(int fs, const ftl::codecs::Channels<0> &s, bool make) {
 	UNIQUE_LOCK(mtx_, lk);
-	if (fs < 0 || (!make && fs >= state_.size())) throw ftl::exception("Frameset index out-of-bounds");
-	if (fs >= state_.size()) state_.resize(fs+1);
+	if (fs < 0 || (!make && static_cast<uint32_t>(fs) >= state_.size())) throw ftl::exception("Frameset index out-of-bounds");
+	if (static_cast<uint32_t>(fs) >= state_.size()) state_.resize(fs+1);
 	state_[fs].selected = s;
 }
 
 ftl::codecs::Channels<0> &Stream::available(int fs) {
 	UNIQUE_LOCK(mtx_, lk);
 	if (fs < 0) throw ftl::exception("Frameset index out-of-bounds");
-	if (fs >= state_.size()) state_.resize(fs+1);
+	if (static_cast<uint32_t>(fs) >= state_.size()) state_.resize(fs+1);
 	return state_[fs].available;
 }
 
@@ -76,7 +76,7 @@ bool Muxer::onPacket(const std::function<void(const ftl::codecs::StreamPacket &,
 }
 
 int Muxer::originStream(int fsid, int fid) {
-	if (fid < revmap_.size()) {
+	if (static_cast<uint32_t>(fid) < revmap_.size()) {
 		return std::get<0>(revmap_[fid]);
 	}
 	return -1;
@@ -135,11 +135,11 @@ void Muxer::reset() {
 int Muxer::_lookup(int sid, int ssid) {
 	SHARED_LOCK(mutex_, lk);
 	auto &se = streams_[sid];
-	if (ssid >= se.maps.size()) {
+	if (static_cast<uint32_t>(ssid) >= se.maps.size()) {
 		lk.unlock();
 		{
 			UNIQUE_LOCK(mutex_, lk2);
-			if (ssid >= se.maps.size()) {
+			if (static_cast<uint32_t>(ssid) >= se.maps.size()) {
 				int nid = nid_++;
 				se.maps.push_back(nid);
 				revmap_.push_back({sid,ssid});
